Cuda compiles, not implemented yet

+++ constensor-core/src/dtype/rand.rs
@@ -138,7 +138,7 @@ impl RandDispatch for i64 {
     }
 }
 #[cfg(all(feature = "cuda", feature = "half"))]
-impl RandDispatch for f16 {
+impl RandDispatch for half::f16 {
     fn cuda_fill_with_uniform(_rng: &CudaRng, _slice: &mut CudaSlice<Self>) -> Result<()> {
         crate::bail!(
             "Uniform random fill is not supported for dtype {}",
@@ -158,7 +158,7 @@ impl RandDispatch for f16 {
     }
 }
 #[cfg(all(feature = "cuda", feature = "bfloat"))]
-impl RandDispatch for bf16 {
+impl RandDispatch for half::bf16 {
     fn cuda_fill_with_uniform(_rng: &CudaRng, _slice: &mut CudaSlice<Self>) -> Result<()> {
         crate::bail!(
             "Uniform random fill is not supported for dtype {}",
